pyproject.toml
[project]
name = "tsp-solver-comparison"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fast-tsp>=0.1.4",
    "ortools>=9.12.4544",
    "python-tsp>=0.5.0",
    "pyvroom>=1.14.0",
    "pyvrp>=0.10.1",
    "ruff>=0.11.2",
    "traveling-rustling>=0.1.1",
    "elkai>=0.2.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "tsplib95>=0.7.0",
    "seaborn>=0.13.2",
]
